pyproject.toml
[project]
name = "monte-carlo"
version = "0.1.0"
description = "Monte Carlo surface tiling and HOOMD-based compression workflows."
readme = "SHAPE_GUIDE.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "matplotlib>=3.7",
    "plotly>=5.18",
    "tqdm>=4.66",
    "scipy>=1.10",
    "gsd>=3.0",
    "pillow>=10.0",
]

[build-system]
requires = ["setuptools>=67"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
